[next Error] type과 다르게 enum은 page에 설정 할 수 없음
