+++ src/app/(template-result)/components/CommentContainer.tsx
@@ -4,8 +4,7 @@ import { Dispatch, SetStateAction } from "react";
 import Button from "@/components/ui/button/Button";
 import CommentEditor from "@/app/(template-result)/components/CommentEditor";
 // import ReplyIcon from "/public/asset/icon/reply.svg";
-import { CommentReponse } from "@/types/comment.type";
-import { COMMENT_EDITOR_TYPE } from "../result/survey/[id]/page";
+import { COMMENT_EDITOR_TYPE, CommentReponse } from "@/types/comment.type";
 
 export default function CommentContainer({
   touchIdx,
